# stats/target/classes/openapi.yaml
openapi: 3.1.0
info:
  title: MetaMapa - Servicio de Estadísticas
  version: 1.0.0
  description: >
    Microservicio de solo lectura que calcula y expone estadísticas agregadas
    (provincias, categorías, distribución horaria y spam) y permite exportarlas en CSV.
    Los datos se recalculan periódicamente mediante un scheduler y se persisten en el
    repositorio local del servicio.
servers:
  - url: https://api.tu-dominio.org/api
    description: Producción
  - url: https://staging.tu-dominio.org/api
    description: Staging
...
security:
  - ApiKeyAuth: []
  # - BearerAuth: []  # Si preferís JWT, habilitá esta línea y deshabilitá ApiKeyAuth arriba.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseProvincia:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/ProvinciaStatsDTO' }
        message: { type: string }
    ApiResponseCategoria:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/CategoriaStatsDTO' }
        message: { type: string }
    ApiResponseSpam:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/SpamStatsDTO' }
        message: { type: string }
    ApiResponseHoraDistribucion:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            categoria: { type: string }
            distribucion:
              type: array
              items: { $ref: '#/components/schemas/HoraDistribucionItem' }
            hora_peak:
              type: integer
              minimum: 0
              maximum: 23
            fechaCalculo:
              type: string
              format: date-time
        message: { type: string }
    ApiResponseDashboard:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          additionalProperties: true
          description: Objeto con los widgets/tiles del dashboard (top provincia, top categoría, hora pico, spam, etc.)
        message: { type: string }
...
    ProvinciaStatsDTO:
      type: object
      properties:
        provincia: { type: string, example: "Córdoba" }
        cantidadHechos: { type: integer, format: int64, example: 532 }
        porcentaje: { type: number, format: double, example: 41.2 }
        fechaCalculo: { type: string, format: date-time, example: "2025-10-30T11:00:00Z" }
        coleccionId: { type: integer, format: int64, example: 123 }
        categoria: { type: string, nullable: true }
    CategoriaStatsDTO:
      type: object
      properties:
        categoria: { type: string, example: "Incendio Forestal" }
        cantidadHechos: { type: integer, format: int64, example: 1220 }
        porcentaje: { type: number, format: double, example: 32.8 }
        fechaCalculo: { type: string, format: date-time, example: "2025-10-30T11:00:00Z" }
    HoraDistribucionItem:
      type: object
      properties:
        hora:
          type: integer
          minimum: 0
          maximum: 23
          example: 19
        cantidadHechos:
          type: integer
          format: int64
          example: 87
        porcentaje:
          type: number
          format: double
          example: 8.6
    SpamStatsDTO:
      type: object
      properties:
        totalSolicitudes: { type: integer, format: int64, example: 1240 }
        solicitudesSpam: { type: integer, format: int64, example: 37 }
        solicitudesNoSpam: { type: integer, format: int64, example: 1203 }
        porcentajeSpam: { type: number, format: double, example: 2.98 }
        porcentajeNoSpam: { type: number, format: double, example: 97.02 }
        fechaCalculo: { type: string, format: date-time, example: "2025-10-30T11:00:00Z" }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Bad Request" }
        message: { type: string, example: "Parámetro 'categoria' es requerido" }
        timestamp: { type: string, format: date-time }
        path: { type: string, example: "/api/stats/hora/categoria/Incendio%20Forestal" }
  parameters:
    ColeccionIdPath:
      name: coleccionId
      in: path
      required: true
      description: Identificador numérico de la colección
      schema: { type: integer, format: int64 }
    CategoriaPath:
      name: categoria
      in: path
      required: true
      description: Nombre de categoría normalizada
      schema: { type: string }
  responses:
    BadRequest:
      description: Parámetros inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
paths:
...
  /stats/provincia/{coleccionId}:
    get:
      tags: [Stats]
      summary: Provincia con más hechos para una colección
      parameters:
        - $ref: '#/components/parameters/ColeccionIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseProvincia' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      provincia: "Córdoba"
                      cantidadHechos: 532
                      porcentaje: 41.2
                      fechaCalculo: "2025-10-30T11:00:00Z"
                      coleccionId: 123
                      categoria: null
                    message: "Provincia con mayor cantidad de hechos"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/categoria:
    get:
      tags: [Stats]
      summary: Categoría con más hechos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseCategoria' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      categoria: "Incendio Forestal"
                      cantidadHechos: 1220
                      porcentaje: 32.8
                      fechaCalculo: "2025-10-30T11:00:00Z"
                    message: "Categoría con mayor cantidad de hechos"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/provincia/categoria/{categoria}:
    get:
      tags: [Stats]
      summary: Provincia top para una categoría
      parameters:
        - $ref: '#/components/parameters/CategoriaPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseProvincia' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      provincia: "Córdoba"
                      cantidadHechos: 311
                      porcentaje: 28.3
                      fechaCalculo: "2025-10-30T11:00:00Z"
                      coleccionId: 123
                      categoria: "Incendio Forestal"
                    message: "Provincia top para la categoría solicitada"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/hora/categoria/{categoria}:
    get:
      tags: [Stats]
      summary: Distribución horaria para una categoría
      parameters:
        - $ref: '#/components/parameters/CategoriaPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseHoraDistribucion' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      categoria: "Incendio Forestal"
                      distribucion:
                        - { hora: 0, cantidadHechos: 8, porcentaje: 0.8 }
                        - { hora: 1, cantidadHechos: 5, porcentaje: 0.5 }
                        - { hora: 19, cantidadHechos: 87, porcentaje: 8.6 }
                        - { hora: 23, cantidadHechos: 17, porcentaje: 1.7 }
                      hora_peak: 19
                      fechaCalculo: "2025-10-30T11:00:00Z"
                    message: "Distribución por horas"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/spam:
    get:
      tags: [Stats]
      summary: Porcentaje y totales de solicitudes de eliminación marcadas como spam
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSpam' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      totalSolicitudes: 1240
                      solicitudesSpam: 37
                      solicitudesNoSpam: 1203
                      porcentajeSpam: 2.98
                      porcentajeNoSpam: 97.02
                      fechaCalculo: "2025-10-30T11:00:00Z"
                    message: "Indicadores de spam"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/dashboard:
    get:
      tags: [Stats]
      summary: Resumen para dashboard de estadísticas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseDashboard' }
              examples:
                ejemplo:
                  value:
                    success: true
                    data:
                      provinciaTop: { provincia: "Córdoba", cantidadHechos: 532 }
                      categoriaTop: { categoria: "Incendio Forestal", cantidadHechos: 1220 }
                      horaPeak: 19
                      spam:
                        porcentajeSpam: 2.98
                        totalSolicitudes: 1240
                      generadoEn: "2025-10-30T11:00:00Z"
                    message: "Dashboard de última corrida"
        '500': { $ref: '#/components/responses/InternalError' }

  # ====================== EXPORT CSV ======================
...